requirements.txt
django
django-allauth
django-bootstrap-v5
django-mptt
